Unit tests for ChartviewController are redundant. Moving tests to ISS_BarometerTests
